platform: move internal headers to internal/
They belong to internal folder to follow our guideline, not in source as they were.
