+++ platform/include/platform/internal/SysTimer.h
