+++ platform/include/platform/internal/mbed_error_hist.h
