+++ platform/include/platform/internal/mbed_os_timer.h
@@ -18,7 +18,7 @@
 #define MBED_MBED_SLEEP_TIMER_H
 
 #include <chrono>
-#include "source/SysTimer.h"
+#include "platform/internal/SysTimer.h"
 
 #if MBED_CONF_RTOS_PRESENT
 extern "C" {
