+++ platform/source/mbed_os_timer.cpp
@@ -17,8 +17,8 @@
 
 #include "platform/mbed_power_mgmt.h"
 #include "platform/CriticalSectionLock.h"
-#include "source/SysTimer.h"
-#include "source/mbed_os_timer.h"
+#include "paltform/internal/SysTimer.h"
+#include "platform/internal/mbed_os_timer.h"
 #include "us_ticker_api.h"
 #include "lp_ticker_api.h"
 #include "mbed_critical.h"
