+++ rtos/source/Semaphore.cpp
@@ -25,7 +25,7 @@
 #include "platform/mbed_assert.h"
 #include "platform/mbed_critical.h"
 #include "platform/mbed_error.h"
-#include "platform/source/mbed_os_timer.h"
+#include "platform/internal/mbed_os_timer.h"
 
 #include <string.h>
