+++ rtos/source/TARGET_CORTEX/mbed_rtx_idle.cpp
@@ -24,7 +24,7 @@
 #include "rtos/source/rtos_idle.h"
 #include "rtos/Kernel.h"
 #include "platform/mbed_power_mgmt.h"
-#include "platform/source/mbed_os_timer.h"
+#include "platform/internal/mbed_os_timer.h"
 #include "TimerEvent.h"
 #include "mbed_critical.h"
 #include "mbed_assert.h"
